extract building blocks into seperate components
