+++ app/components/building-blocks/Date.tsx
@@ -1,0 +1,8 @@
+import { useTicket } from '@/app/TicketContext'
+import moment from 'moment'
+
+export default function Date() {
+    const { data } = useTicket()
+
+    return <p>{moment(data.datetime).format('dddd, DD. MMM YYYY, HH:mm')}</p>
+}
