+++ app/components/building-blocks/Seating.tsx
@@ -1,0 +1,7 @@
+import { useTicket } from '@/app/TicketContext'
+
+export default function Seating() {
+    const { data } = useTicket()
+
+    return <p>{data.seatType}</p>
+}
