+++ app/components/building-blocks/TicketBarcode.tsx
